se intentó corregir error de footer pero sin éxito
