+++ src/components/footer.jsx
@@ -1,78 +1,111 @@
-import '../App.css';
 
-export default function Footer() {
+import '../styles/footer.css';
+/*  import { Link } from 'react-router-dom';
+import logospafooter from '../assets/Img/logospafooter.png';
+import logomastercard from '../assets/Img/logomastercard.png';  */
+
+function FooterComponent (){
   return (
-    <footer className="row row-cols-1 row-cols-md-5 py-5 my-5 border-top">
-      <div className="col mb-3">
-        <a href="/" className="d-flex align-items-center mb-3 link-body-emphasis text-decoration-none" aria-label="Bootstrap">
-        </a>
-        <p className="text-body-secondary">© 2025</p>
-      </div>
-
-      <div className="col mb-3">
-        <h5>Section</h5>
-        <ul className="nav flex-column">
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Home</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Features</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Pricing</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">FAQs</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">About</a>
-          </li>
-        </ul>
-      </div>
-
-      <div className="col mb-3">
-        <h5>Section</h5>
-        <ul className="nav flex-column">
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Home</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Features</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Pricing</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">FAQs</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">About</a>
-          </li>
-        </ul>
-      </div>
-
-      <div className="col mb-3">
-        <h5>Section</h5>
-        <ul className="nav flex-column">
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Home</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Features</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">Pricing</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">FAQs</a>
-          </li>
-          <li className="nav-item mb-2">
-            <a href="#" className="nav-link p-0 text-body-secondary">About</a>
-          </li>
-        </ul>
-      </div>
-    </footer>
-  );
+    <>
+      {/* <footer className="row row-cols-1 row-cols-sm-2 row-cols-md-5 py-5 py-3 border-top"> 
+        <div className="card mb-3" >
+          <div className="row g-0">
+            <div className="col-md-4">
+              <img src={logospafooter} className="img-fluid rounded-start" alt="Logo-Nirvana-Spa-&-Beauty-footer" />
+            </div>
+            <div className="col-md-8">
+              <div className="card-body">
+                <h5 className="card-title">SÍGUENOS</h5>
+                <div className="RedesIMG">
+                  <a href="https://www.facebook.com/" className="redes"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-facebook" viewBox="0 0 16 16"><path d="M16 8.049c0-4.446-3.582-8.05-8-8.05C3.58 0-.002 3.603-.002 8.05c0 4.017 2.926 7.347 6.75 7.951v-5.625h-2.03V8.05H6.75V6.275c0-2.017 1.195-3.131 3.022-3.131.876 0 1.791.157 1.791.157v1.98h-1.009c-.993 0-1.303.621-1.303 1.258v1.51h2.218l-.354 2.326H9.25V16c3.824-.604 6.75-3.934 6.75-7.951"/></svg></a>
+                  <a href="https://www.instagram.com/" className="redes"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-instagram" viewBox="0 0 16 16"><path d="M8 0C5.829 0 5.556.01 4.703.048 3.85.088 3.269.222 2.76.42a3.9 3.9 0 0 0-1.417.923A3.9 3.9 0 0 0 .42 2.76C.222 3.268.087 3.85.048 4.7.01 5.555 0 5.827 0 8.001c0 2.172.01 2.444.048 3.297.04.852.174 1.433.372 1.942.205.526.478.972.923 1.417.444.445.89.719 1.416.923.51.198 1.09.333 1.942.372C5.555 15.99 5.827 16 8 16s2.444-.01 3.298-.048c.851-.04 1.434-.174 1.943-.372a3.9 3.9 0 0 0 1.416-.923c.445-.445.718-.891.923-1.417.197-.509.332-1.09.372-1.942C15.99 10.445 16 10.173 16 8s-.01-2.445-.048-3.299c-.04-.851-.175-1.433-.372-1.941a3.9 3.9 0 0 0-.923-1.417A3.9 3.9 0 0 0 13.24.42c-.51-.198-1.092-.333-1.943-.372C10.443.01 10.172 0 7.998 0zm-.717 1.442h.718c2.136 0 2.389.007 3.232.046.78.035 1.204.166 1.486.275.373.145.64.319.92.599s.453.546.598.92c.11.281.24.705.275 1.485.039.843.047 1.096.047 3.231s-.008 2.389-.047 3.232c-.035.78-.166 1.203-.275 1.485a2.5 2.5 0 0 1-.599.919c-.28.28-.546.453-.92.598-.28.11-.704.24-1.485.276-.843.038-1.096.047-3.232.047s-2.39-.009-3.233-.047c-.78-.036-1.203-.166-1.485-.276a2.5 2.5 0 0 1-.92-.598 2.5 2.5 0 0 1-.6-.92c-.109-.281-.24-.705-.275-1.485-.038-.843-.046-1.096-.046-3.233s.008-2.388.046-3.231c.036-.78.166-1.204.276-1.486.145-.373.319-.64.599-.92s.546-.453.92-.598c.282-.11.705-.24 1.485-.276.738-.034 1.024-.044 2.515-.045zm4.988 1.328a.96.96 0 1 0 0 1.92.96.96 0 0 0 0-1.92m-4.27 1.122a4.109 4.109 0 1 0 0 8.217 4.109 4.109 0 0 0 0-8.217m0 1.441a2.667 2.667 0 1 1 0 5.334 2.667 2.667 0 0 1 0-5.334"/></svg></a>
+                  <a href="https://www.tiktok.com/" className="redes"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-tiktok" viewBox="0 0 16 16"><path d="M9 0h1.98c.144.715.54 1.617 1.235 2.512C12.895 3.389 13.797 4 15 4v2c-1.753 0-3.07-.814-4-1.829V11a5 5 0 1 1-5-5v2a3 3 0 1 0 3 3z"/></svg></a>
+                  <a href="https://x.com/" className="redes"><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-twitter-x" viewBox="0 0 16 16"><path d="M12.6.75h2.454l-5.36 6.142L16 15.25h-4.937l-3.867-5.07-4.425 5.07H.316l5.733-6.57L0 .75h5.063l3.495 4.633L12.601.75Zm-.86 13.028h1.36L4.323 2.145H2.865z"/></svg></a>
+                </div>
+              </div>
+              
+            </div>
+          </div>
+        </div>
+        <div className="columna2">
+          <div className="col mb-3"> 
+              <h5>NUESTROS DATOS</h5> 
+              <div className="textofootercol">
+                <ul className="nav flex-column"> 
+                  <li className="nav-item mb-2">
+                    <Link to="*" className="nav-link p-0 text-body-secondary"><p>Atención al Cliente - 0381 5783 030</p></Link>
+                  </li> 
+                  <li className="nav-item mb-2">
+                    <Link to="*" className="nav-link p-0 text-body-secondary"><p>Trabaja con Nosotros 
+                      <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-linkedin" viewBox="0 0 16 16"><path d="M0 1.146C0 .513.526 0 1.175 0h13.65C15.474 0 16 .513 16 1.146v13.708c0 .633-.526 1.146-1.175 1.146H1.175C.526 16 0 15.487 0 14.854zm4.943 12.248V6.169H2.542v7.225zm-1.2-8.212c.837 0 1.358-.554 1.358-1.248-.015-.709-.52-1.248-1.342-1.248S2.4 3.226 2.4 3.934c0 .694.521 1.248 1.327 1.248zm4.908 8.212V9.359c0-.216.016-.432.08-.586.173-.431.568-.878 1.232-.878.869 0 1.216.662 1.216 1.634v3.865h2.401V9.25c0-2.22-1.184-3.252-2.764-3.252-1.274 0-1.845.7-2.165 1.193v.025h-.016l.016-.025V6.169h-2.4c.03.678 0 7.225 0 7.225z"/></svg></p>
+                    </Link>
+                  </li> 
+                  <li className="nav-item mb-2">
+                    <Link to="/contacto" className="nav-link p-0 text-body-secondary"><p>Contáctanos</p></Link>
+                  </li> 
+                  <li className="nav-item mb-2">
+                    <Link to="/nosotros" className="nav-link p-0 text-body-secondary"><p>Nosotros</p></Link>
+                  </li>
+                  <li className="nav-item mb-2">
+                    <Link href="*" className="nav-link p-0 text-body-secondary"><p>ChatBot</p></Link>
+                  </li>
+                  <li className="nav-item mb-2">
+                    <a href="https://maps.app.goo.gl/LiMqUe2PFhkN5obY9" target="_blank" className="nav-link p-0 text-body-secondary">
+                      <p><svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-geo-alt-fill" viewBox="16 16">
+                      <path d="M8 16s6-5.686 6-10A6 6 0 0 0 2 6c0 4.314 6 10 6 10m0-7a3 3 0 1 1 0-6 3 3 0 0 1 0 6"/>
+                      </svg>Gral. José María Paz 576, <br/>
+                      San Miguel de Tucumán, Tucumán</p>
+                    </a>
+                  </li>
+                </ul>
+              </div>
+            </div>
+        </div>  
+        <div className="columna3">
+              <div className="col mb-3">
+                <h5>INFORMACIÓN DE LA EMPRESA</h5> 
+                <div className="textofootercol">
+                  <ul className="nav flex-column"> 
+                    <li className="nav-item mb-2"><Link href="*" className="nav-link p-0 text-body-secondary"><p>Medios de Pago:</p></Link>
+                    </li>
+                    <li className="nav-item mb-2">
+                      <Link href="./Pages/PaginaError.html" className="nav-link p-0 text-body-secondary">
+                      
+                        <img src={logomastercard} alt="logo-mastercard" className="mediopago" width="30" />
+                      
+                        <svg xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" width="50" viewBox="0 0 48 48" class="mediopago"><path d="M-.003 36.119V12.889c0-2.697 2.191-4.888 4.888-4.888H43.11c2.697 0 4.888 2.191 4.888 4.888v23.231c0 2.697-2.191 4.888-4.888 4.888H4.885C2.187 41.007-.003 38.817-.003 36.119zM12.134 31.001l5.271-13h-3.416l-3.387 8.845-1.369-7.52c-.16-.841-.795-1.325-1.5-1.325H2.198l-.076.376c1.136.256 2.428.668 3.208 1.107.48.268.616.501.773 1.143l2.593 10.375H12.134zM19.363 31.001l2.688-13h-3.251l-2.692 13H19.363zM34.122 31.001l.7-1.999h4.14l.404 1.999h3.007l-2.625-13h-2.776c-.625 0-1.152.375-1.385.955l-4.88 12.045H34.122zM30.875 26.675c.015-3.308-4.587-3.491-4.556-4.969.009-.452.44-.929 1.379-1.049.468-.06 1.751-.109 3.208.56l.571-2.66c-.783-.283-1.791-.556-3.041-.556-3.216 0-5.477 1.704-5.497 4.148-.021 1.805 1.615 2.811 2.847 3.413 1.267.613 1.693 1.007 1.689 1.557-.009.841-1.012 1.212-1.948 1.228-1.636.023-2.584-.443-3.34-.792l-.589 2.747c.76.348 2.163.651 3.617.667C28.63 30.969 30.866 29.285 30.875 26.675zM35.875 26.001l1.573-4.487.907 4.487H35.875zM8.123 10.126h-3c-1.808 0-3 1.443-3 3v1C3.374 11.95 5.578 10.847 8.123 10.126zM23.998 47.001c11.04 0 20 .225 20 .501s-8.96.501-20 .501-20-.225-20-.501S12.958 47.001 23.998 47.001z"></path></svg>
+                      
+                        <svg xmlns="http://www.w3.org/2000/svg" width="30" height="25" fill="currentColor" class="bi bi-paypal" viewBox="0 0 16 16"><path d="M14.06 3.713c.12-1.071-.093-1.832-.702-2.526C12.628.356 11.312 0 9.626 0H4.734a.7.7 0 0 0-.691.59L2.005 13.509a.42.42 0 0 0 .415.486h2.756l-.202 1.28a.628.628 0 0 0 .62.726H8.14c.429 0 .793-.31.862-.731l.025-.13.48-3.043.03-.164.001-.007a.35.35 0 0 1 .348-.297h.38c1.266 0 2.425-.256 3.345-.91q.57-.403.993-1.005a4.94 4.94 0 0 0 .88-2.195c.242-1.246.13-2.356-.57-3.154a2.7 2.7 0 0 0-.76-.59l-.094-.061ZM6.543 8.82a.7.7 0 0 1 .321-.079H8.3c2.82 0 5.027-1.144 5.672-4.456l.003-.016q.326.186.548.438c.546.623.679 1.535.45 2.71-.272 1.397-.866 2.307-1.663 2.874-.802.57-1.842.815-3.043.815h-.38a.87.87 0 0 0-.863.734l-.03.164-.48 3.043-.024.13-.001.004a.35.35 0 0 1-.348.296H5.595a.106.106 0 0 1-.105-.123l.208-1.32z"/></svg>
+                  
+                        <svg xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" width="0"  viewBox="0 0 48 48" class="mediopago"><ellipse cx="23.5" cy="23.5" fill="#4fc3f7" rx="21.5" ry="15.5"></ellipse><path fill="#fafafa" d="M22.471,24.946c-1.978-5.537-4.884-10.881-6.085-12.995c-0.352-0.619-0.787-1.186-1.29-1.69 l-2.553-2.553c-0.391-0.391-1.414,0-1.414,0L9.497,8.734l-0.162,2.319L8.773,11c-0.518,0-0.938,0.42-0.938,0.938 c0,0.52,0.413,0.969,0.933,0.961c1.908-0.03,3.567,1.601,3.567,1.601h2c0.32,0.32,1.139,1.366,1.328,2.439 c0.107,0.611,0.154,1.229,0.119,1.848C15.458,24.622,16.835,26,16.835,26c-5.5-3.5-14.819-2.964-14.819-2.964l0.193,3.016L5,31 c0.919,0.212,0.744-0.626,1.765-0.504c6.199,0.741,13.57,0.004,13.57,0.004c1.5,0,1.958-0.793,2.665-1.5 C24,28,22.849,26.004,22.471,24.946z"></path><path fill="#fafafa" d="M24.913,24.946c1.978-5.537,4.884-10.881,6.085-12.995c0.352-0.619,0.787-1.186,1.29-1.69 l2.553-2.553c0.391-0.391,1.414,0,1.414,0L37.814,9l0.235,2.053L38.611,11c0.518,0,0.938,0.42,0.938,0.938 c0,0.52-0.413,0.969-0.933,0.961c-1.908-0.03-3.567,1.601-3.567,1.601h-2c-0.32,0.32-1.139,1.366-1.328,2.439 c-0.107,0.611-0.154,1.229-0.119,1.848C31.926,24.622,30.549,26,30.549,26c5.5-3.5,15-3,15-3l-0.165,3l-3,5 c-0.919,0.212-0.744-0.626-1.765-0.504c-6.199,0.741-13.57,0.004-13.57,0.004c-1.5,0-1.958-0.793-2.665-1.5 C23.384,28,24.535,26.004,24.913,24.946z"></path><path fill="#1a237e" d="M43.832,16.326c-0.311-0.415-0.644-0.808-0.992-1.187c-0.059-0.064-0.123-0.123-0.183-0.186 c-0.309-0.326-0.628-0.639-0.96-0.938c-0.026-0.023-0.053-0.045-0.079-0.068c-0.587-0.522-1.201-1.012-1.845-1.454 c0.071-0.175,0.11-0.364,0.11-0.555c0-0.792-0.643-1.437-1.481-1.437c-0.001,0-0.003,0-0.004,0l-0.015,0.002V9.32 c0-0.534-0.288-1.032-0.75-1.299L36.269,7.24c-0.221-0.085-1.356-0.478-1.946,0.113l-1.837,1.838 c-0.381-0.106-0.89-0.25-1.211-0.326C28.893,8.288,26.446,8.014,24,8c-3.031-0.004-6.095,0.39-9.018,1.275l-1.921-1.921 c-0.59-0.59-1.725-0.199-2.018-0.079L9.75,8.021C9.288,8.288,9,8.786,9,9.32v1.186L8.938,10.5c-0.793,0-1.438,0.646-1.438,1.438 c0,0.311,0.103,0.614,0.283,0.865c-0.978,0.715-1.903,1.512-2.722,2.422c-0.315,0.35-0.616,0.715-0.9,1.096 C2.638,18.346,2.061,20.87,2,23.5c-0.035,2.628,0.455,5.223,1.932,7.343c1.478,2.132,3.451,3.854,5.624,5.163 c4.378,2.609,9.436,3.749,14.444,3.846c2.511-0.026,5.023-0.319,7.471-0.924c2.442-0.624,4.81-1.582,6.986-2.9 c2.163-1.328,4.143-3.041,5.617-5.18c1.476-2.122,1.932-4.719,1.894-7.347C45.905,20.87,45.357,18.348,43.832,16.326z M40.793,15.139c0.229,0.225,0.448,0.459,0.662,0.697c0.096,0.107,0.195,0.211,0.288,0.32c0.293,0.347,0.573,0.703,0.828,1.076 c1.088,1.579,1.785,3.39,1.957,5.242c-2.274-0.031-8.444,0.114-13.042,2.342c0.335-1.133,0.619-3.016,0.449-6.058 c-0.03-0.552,0.008-1.135,0.113-1.733c0.139-0.79,0.702-1.618,1.054-2.026h0.727c0.731,0,1.432-0.224,2.025-0.647 c0.624-0.444,1.559-0.981,2.588-0.954c0.072,0,0.139-0.03,0.21-0.04c0.267,0.192,0.536,0.383,0.792,0.587 c0.076,0.061,0.15,0.124,0.225,0.186c0.273,0.224,0.538,0.457,0.795,0.696C40.576,14.93,40.686,15.034,40.793,15.139z M24,9 c2.369,0.026,4.734,0.303,7.027,0.87c0.208,0.053,0.412,0.118,0.617,0.181c-0.482,0.503-0.906,1.054-1.246,1.652 c-1.175,2.068-4.124,7.483-6.121,13.075c-0.075,0.208-0.163,0.43-0.255,0.66c-0.112,0.281-0.226,0.572-0.331,0.868 c-0.104-0.296-0.219-0.588-0.331-0.868c-0.092-0.23-0.18-0.452-0.255-0.66c-2-5.599-4.947-11.009-6.121-13.075 c-0.297-0.523-0.667-1.004-1.074-1.456C18.522,9.461,21.264,9.054,24,9z M5.435,17.238c0.251-0.364,0.524-0.713,0.811-1.052 c0.094-0.112,0.196-0.218,0.294-0.327c0.202-0.225,0.408-0.448,0.625-0.662c0.115-0.114,0.233-0.224,0.351-0.335 c0.229-0.213,0.463-0.421,0.704-0.622c0.099-0.083,0.198-0.166,0.299-0.247c0.243-0.193,0.495-0.376,0.748-0.558 c0.886,0.089,1.707,0.522,2.262,0.918C12.123,14.776,12.823,15,13.555,15h0.727c0.352,0.407,0.915,1.235,1.054,2.026 c0.105,0.597,0.143,1.18,0.113,1.733c-0.17,3.042,0.114,4.927,0.449,6.059c-4.193-2.029-9.734-2.333-12.425-2.344 C3.648,20.623,4.346,18.814,5.435,17.238z M6.236,30.271c-0.192-0.224-0.396-0.437-0.572-0.673 C4.329,27.826,3.49,25.705,3.426,23.5c0-0.008,0.001-0.017,0.001-0.025c2.878,0.006,9.226,0.351,13.305,2.947 c0.211,0.134,0.484,0.088,0.646-0.104c0.162-0.19,0.153-0.477-0.014-0.662c-0.012-0.014-1.218-1.422-0.916-6.842 c0.035-0.63-0.007-1.29-0.126-1.962c-0.218-1.235-1.133-2.372-1.467-2.706C14.76,14.053,14.632,14,14.5,14h-0.945 c-0.522,0-1.021-0.159-1.445-0.462c-0.745-0.531-1.925-1.147-3.185-1.14c-0.131,0.004-0.226-0.063-0.281-0.117 C8.552,12.192,8.5,12.067,8.5,11.938c0-0.242,0.196-0.438,0.391-0.44l0.562,0.054c0.111,0.007,0.216-0.027,0.308-0.084l0.386,0.386 C10.242,11.949,10.37,12,10.5,12c0.053,0,0.106-0.009,0.158-0.025l1.207-0.402l1.281,1.281C13.244,12.951,13.372,13,13.5,13 s0.256-0.049,0.354-0.146c0.195-0.195,0.195-0.512,0-0.707L12.707,11l0.146-0.146C12.951,10.756,13,10.628,13,10.5 s-0.049-0.256-0.146-0.354l-1-1c-0.195-0.195-0.512-0.195-0.707,0C11.049,9.244,11,9.372,11,9.5s0.049,0.256,0.146,0.354 l0.646,0.646l-0.063,0.063l-1.095,0.365L10,10.293V9.32c0-0.178,0.096-0.344,0.25-0.434l1.22-0.712 c0.365-0.139,0.792-0.179,0.883-0.114l2.554,2.554c0.475,0.475,0.882,1.007,1.209,1.583c1.161,2.043,4.076,7.393,6.049,12.917 c0.078,0.219,0.171,0.452,0.267,0.694c0.347,0.871,0.741,1.858,0.58,2.583C22.808,29.309,21.728,30,20.49,30 c-0.07,0.002-7.123,0.139-13.425,0.011C6.798,30.002,6.509,30.114,6.236,30.271z M37.217,33.918 c-1.98,1.119-4.156,1.898-6.385,2.419c-2.228,0.539-4.528,0.798-6.832,0.812c-4.592,0.01-9.259-0.951-13.23-3.208 c-1.401-0.799-2.709-1.764-3.832-2.891c0.036-0.014,0.083-0.038,0.107-0.039C13.367,31.138,20.439,31.001,20.5,31 c1.396,0,2.616-0.673,3.192-1.67c0.575,0.997,1.794,1.67,3.182,1.67c0.071,0.002,7.146,0.139,13.462,0.011 c0.089,0.003,0.272,0.102,0.483,0.249C39.748,32.289,38.531,33.185,37.217,33.918z M42.329,29.593 c-0.247,0.329-0.526,0.635-0.803,0.941c-0.37-0.273-0.81-0.524-1.192-0.524c-0.005,0-0.011,0-0.017,0 c-6.3,0.125-13.354-0.01-13.434-0.011c-1.228,0-2.308-0.691-2.512-1.608c-0.161-0.725,0.232-1.712,0.58-2.583 c0.096-0.242,0.189-0.476,0.267-0.694c1.971-5.518,4.887-10.871,6.049-12.917c0.327-0.576,0.734-1.108,1.209-1.583l2.55-2.551 C35.122,8,35.548,8.037,35.841,8.14l1.293,0.747c0.154,0.09,0.25,0.256,0.25,0.434v0.973l-0.635,0.635l-1.095-0.365L35.591,10.5 l0.646-0.646c0.098-0.098,0.146-0.226,0.146-0.354s-0.049-0.256-0.146-0.354c-0.195-0.195-0.512-0.195-0.707,0l-1,1 c-0.098,0.098-0.146,0.226-0.146,0.354s0.049,0.256,0.146,0.354L34.677,11l-1.146,1.146c-0.195,0.195-0.195,0.512,0,0.707 C33.628,12.951,33.756,13,33.884,13s0.256-0.049,0.354-0.146l1.281-1.281l1.207,0.402C36.777,11.991,36.831,12,36.884,12 c0.13,0,0.258-0.051,0.354-0.146l0.386-0.386c0.092,0.057,0.197,0.092,0.308,0.084l0.515-0.052c0.242,0,0.438,0.196,0.438,0.438 c0,0.129-0.052,0.254-0.143,0.343c-0.056,0.055-0.157,0.109-0.282,0.117c-1.279,0.011-2.439,0.608-3.185,1.14 C34.851,13.841,34.352,14,33.83,14h-0.946c-0.133,0-0.26,0.053-0.354,0.146c-0.334,0.334-1.25,1.473-1.467,2.706 c-0.118,0.674-0.161,1.334-0.126,1.963c0.302,5.419-0.904,6.827-0.907,6.831c-0.18,0.181-0.196,0.468-0.037,0.666 c0.159,0.199,0.442,0.246,0.659,0.109c4.408-2.805,11.576-2.969,13.922-2.942c0,0.007,0.001,0.013,0.001,0.02 C44.507,25.705,43.666,27.824,42.329,29.593z"></path></svg>
+                      </Link>
+                    </li>         
+                    <li className="nav-item mb-2">
+                      <Link to="*" className="nav-link p-0 text-body-secondary"><p>Centro de Ayuda</p></Link>
+                    </li> 
+                    <li class="nav-item mb-2">
+                      <Link to="*" className="nav-link p-0 text-body-secondary"><p>Cancelar un turno</p></Link>
+                    </li> 
+                    <li class="nav-item mb-2">
+                      <Link to="*" className="nav-link p-0 text-body-secondary"><p>Política de Privacidad</p></Link>
+                    </li>  
+                  </ul>
+              </div>        
+            </div> 
+        </div>
+        <div className="DownText">
+          <p> Todos los derechos reservados   |   ® Nirvana Spa & Beauty Company 2025</p>      
+        </div>
+        
+        <div className="wp cajawp">
+          <a href="https://web.whatsapp.com/" className="visitado" target="_blank"><i className="bi bi-whatsapp"></i></a>
+        </div>
+      </footer>  */}
+      <footer className='footer-content'>
+        <p>este es el footer</p>
+      </footer>
+    </>
+  )  
 }
 
-
+export default FooterComponent;
